Move test_positive_install_sat_with_katello_certs to non-destructive
